+++ src/browser/index.js
@@ -1,0 +1,1 @@
+export { decodeBrowserImage } from './decode.js';